requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-filter
